Changed button from folder lib to src and followed index exportation pattern
