+++ src/components/Button/styles.scss
@@ -1,81 +1,69 @@
-// styles.scss simplificado (apenas light)
-.button {
-  display: flex;
+.btn {
+  display: inline-flex;
   align-items: center;
   justify-content: center;
-  min-width: 40px;
-  gap: 8px;
-  padding: 10px;
-  border: 0;
-  border-radius: 4px;
-  cursor: pointer;
+  border-radius: $spacing-sm;
+  font-weight: 500;
   font-family: $font-family-default;
-  font-size: $font-size-small;
-  font-weight: $font-weight-regular;
-  line-height: $line-height-caption;
-  transition: all 0.2s linear;
+  transition: background-color 0.2s ease, border-color 0.2s ease, color 0.2s ease;
+  cursor: pointer;
 
-  &--primary {
-    color: $background-base;
-    background-color: $color-primary;
+  &:focus-visible {
+    outline: none;
+    box-shadow: 0 0 0 2px $color-primary, 0 0 0 4px $background-base;
+  }
 
-    path {
-      fill: $background-base;
-    }
+  &:disabled {
+    opacity: 0.5;
+    pointer-events: none;
+    background: $background-disabled;
+    color: $color-disabled;
+    border-color: $border-disabled;
+  }
 
-    &:hover:not(:disabled) {
-      opacity: 0.9;
-    }
+  &--default {
+    background-color: $background-primary;
+    color: $color-white;
 
-    &:active:not(:disabled) {
-      opacity: 0.8;
-    }
-
-    &:disabled {
-      color: $color-disabled;
-      background-color: $background-disabled;
-      cursor: not-allowed;
-
-      path {
-        fill: $color-disabled;
-      }
+    &:hover {
+      background-color: color-mix(in srgb, $background-primary 90%, $color-base);
     }
   }
 
-  &--secondary {
-    color: $background-base;
-    background-color: $color-secondary;
+  &--outline {
+    border: 1px solid $border-base;
+    color: $color-base;
+    background: transparent;
 
-    path {
-      fill: $background-base;
-    }
-
-    &:hover:not(:disabled) {
-      opacity: 0.9;
-    }
-
-    &:active:not(:disabled) {
-      opacity: 0.8;
-    }
-
-    &:disabled {
-      color: $color-disabled;
-      background-color: $background-disabled;
-      cursor: not-allowed;
-
-      path {
-        fill: $color-disabled;
-      }
+    &:hover {
+      background-color: $color-soft;
     }
   }
 
-  &:focus {
-    outline: 2px solid $color-primary;
-    outline-offset: 2px;
+  &--ghost {
+    background: transparent;
+    color: $color-base;
+
+    &:hover {
+      background-color: $color-soft;
+    }
   }
 
-  @media (max-width: 768px) {
-    padding: 8px 12px;
-    font-size: 13px;
+  &--sm {
+    height: 1.5rem;
+    padding: 1rem 1.5rem;
+    font-size: $font-size-caption;
+  }
+
+  &--md {
+    height: 2.5rem;
+    padding: 1.5rem 2rem;
+    font-size: calc($font-size-body * 1.1);
+  }
+
+  &--lg {
+    height: 3.5rem;
+    padding: 2rem 2.5rem;
+    font-size: calc($font-size-body * 1.2);
   }
 }
