--- src/lib/components/Button/Button.scss
@@ -1,69 +1,0 @@
-.btn {
-  display: inline-flex;
-  align-items: center;
-  justify-content: center;
-  border-radius: $spacing-sm;
-  font-weight: 500;
-  font-family: $font-family-default;
-  transition: background-color 0.2s ease, border-color 0.2s ease, color 0.2s ease;
-  cursor: pointer;
-
-  &:focus-visible {
-    outline: none;
-    box-shadow: 0 0 0 2px $color-primary, 0 0 0 4px $background-base;
-  }
-
-  &:disabled {
-    opacity: 0.5;
-    pointer-events: none;
-    background: $background-disabled;
-    color: $color-disabled;
-    border-color: $border-disabled;
-  }
-
-  &--default {
-    background-color: $background-primary;
-    color: $color-white;
-
-    &:hover {
-      background-color: color-mix(in srgb, $background-primary 90%, $color-base);
-    }
-  }
-
-  &--outline {
-    border: 1px solid $border-base;
-    color: $color-base;
-    background: transparent;
-
-    &:hover {
-      background-color: $color-soft;
-    }
-  }
-
-  &--ghost {
-    background: transparent;
-    color: $color-base;
-
-    &:hover {
-      background-color: $color-soft;
-    }
-  }
-
-  &--sm {
-    height: 1.5rem;
-    padding: 1rem 1.5rem;
-    font-size: $font-size-caption;
-  }
-
-  &--md {
-    height: 2.5rem;
-    padding: 1.5rem 2rem;
-    font-size: calc($font-size-body * 1.1);
-  }
-
-  &--lg {
-    height: 3.5rem;
-    padding: 2rem 2.5rem;
-    font-size: calc($font-size-body * 1.2);
-  }
-}
