+++ src/stories/Button.stories.tsx
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from "@storybook/react-vite";
-import { Button } from "../lib/components/Button/Button";
+import { Button } from "../components/Button";
 
 const meta: Meta<typeof Button> = {
   title: "Components/Button",
